Now cmake options list use consistently capital letter names to match the usual practice.
With cmake compiler warning options are honored (mostly
ignored before)

The *.list files are modified for human readability.

Where useful there are ifdef/define/include to aid Windows builds

Here and there some C statements modified slightly
for greater form consistency.
Some calls have new casts in arguments to eliminate warnings
from certain compilers.

	modified:   CMakeLists.txt

	modified:   appveyor.yml
	modified:   config.h.in
	modified:   configure.ac
	modified:   dwarfdump/CMakeLists.txt

	modified:   dwarfdump/config.h.in.cmake
	modified:   dwarfdump/dwarfdump.c
	modified:   dwarfdump/dwconf.c
	modified:   dwarfdump/esb.c
	modified:   dwarfdump/globals.h
	modified:   dwarfdump/tag_attr.list
	modified:   dwarfdump/tag_attr_ext.list
	modified:   dwarfdump/tag_tree.list
	modified:   dwarfdump/tag_tree_ext.list
	modified:   dwarfexample/CMakeLists.txt

	modified:   dwarfexample/config.h.in.cmake
	modified:   dwarfexample/frame1.c
	modified:   dwarfgen/CMakeLists.txt

	modified:   dwarfgen/config.h.in.cmake
	modified:   dwarfgen/createirepformfrombinary.cc
	modified:   dwarfgen/createirepfrombinary.cc
	modified:   libdwarf/CMakeLists.txt
	modified:   libdwarf/config.h.in.cmake
	modified:   libdwarf/dwarf_elf_access.c
	modified:   libdwarf/dwarf_elf_defines.h
	modified:   libdwarf/dwarf_elf_load_headers.c
	modified:   libdwarf/dwarf_elfread.c
	modified:   libdwarf/dwarf_frame2.c
	modified:   libdwarf/dwarf_incl.h
	modified:   libdwarf/dwarf_machoread.c
	modified:   libdwarf/dwarf_object_detector.c
	modified:   libdwarf/dwarf_object_read_common.c
	modified:   libdwarf/dwarf_peread.c
	modified:   libdwarf/dwarf_query.c
	modified:   libdwarf/gennames.c

A new DW_FTYPE added.
	modified:   libdwarf/generated_libdwarf.h.in
	modified:   libdwarf/libdwarf.h.in

	modified:   libdwarf/pro_encode_nm.c
	modified:   libdwarf/pro_incl.h

Modified the cmake commands to match the new option names.
	modified:   scripts/buildandreleasetest.sh

+++ libdwarf/pro_encode_nm.c
@@ -66,7 +66,7 @@ _dwarf_pro_encode_leb128_nm(Dwarf_Unsigned val, int *nbytes,
         *a = uc;
         a++;
     } while (val);
-    *nbytes = a - space;
+    *nbytes = (int)(a - space);
     return DW_DLV_OK;
 }
 
@@ -105,6 +105,6 @@ _dwarf_pro_encode_signed_leb128_nm(Dwarf_Signed value, int *nbytes,
         *str = byte;
         str++;
     } while (more);
-    *nbytes = str - space;
+    *nbytes = (int)(str - space);
     return DW_DLV_OK;
 }
